# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "ApiResponse[NoneType]",
        "type": "object"
      },
...
      "Body_add_vector_multiple_knowledge_add_multiple_post": {
        "properties": {
          "files": {
            "description": "要上传的文件列表，仅支持PDF和TXT格式",
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_add_vector_multiple_knowledge_add_multiple_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/knowledge/add/multiple": {
      "post": {
        "description": "上传多个文件，将文件保存到向量数据库，仅支持TXT和PDF",
        "operationId": "add_vector_multiple_knowledge_add_multiple_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_vector_multiple_knowledge_add_multiple_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Vector Multiple",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/add/multiple/stream": {
      "post": {
        "description": "上传多个文件，流式返回处理进度，仅支持TXT、PDF、MD、PPTX、DOCX",
        "operationId": "add_vector_multiple_stream_knowledge_add_multiple_stream_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_vector_multiple_stream_knowledge_add_multiple_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"schema_version\":\"1.0\",\"type\":\"response\",\"content\":\"hello\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Versioned Server-Sent Events stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Vector Multiple Stream",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/add/single": {
      "post": {
        "description": "上传文件，将文件保存到向量数据库，仅支持TXT和PDF",
        "operationId": "add_vector_single_knowledge_add_single_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_vector_single_knowledge_add_single_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Vector Single",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/chunks": {
      "get": {
        "description": "获取文档切片信息",
        "operationId": "get_document_chunks_knowledge_chunks_get",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DocumentChunksResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Chunks",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/clean": {
      "delete": {
        "description": "删除用户上传的所有向量",
        "operationId": "clean_user_vectors_knowledge_clean_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clean User Vectors",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/delete/filename": {
      "delete": {
        "description": "通过文件名删除MD5记录及其对应的知识库文档\n:param filename: 要删除的文件名\n:param delete_documents: 是否同时删除知识库文档（默认True）",
        "operationId": "delete_by_filename_knowledge_delete_filename_delete",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "delete_documents",
            "required": false,
            "schema": {
              "default": true,
              "title": "Delete Documents",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete By Filename",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/detail": {
      "get": {
        "description": "获取文档详情内容",
        "operationId": "get_document_detail_knowledge_detail_get",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeDocumentDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Detail",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/embedding/current": {
      "get": {
        "operationId": "get_current_embedding_config_knowledge_embedding_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current Embedding Config",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/embedding/ollama/models": {
      "get": {
        "operationId": "list_embedding_ollama_models_knowledge_embedding_ollama_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "base_url",
            "required": false,
            "schema": {
              "default": "http://localhost:11434",
              "title": "Base Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Embedding Ollama Models",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/embedding/switch": {
      "post": {
        "operationId": "switch_embedding_and_rebuild_knowledge_embedding_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Switch Embedding And Rebuild",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/image/{md5}/{filename}": {
      "get": {
        "description": "返回PDF中提取的原始图片（需JWT鉴权）\n图片存储在 data/extracted_images/{user_id}/{md5}/{filename}",
        "operationId": "serve_knowledge_image_knowledge_image__md5___filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "md5",
            "required": true,
            "schema": {
              "title": "Md5",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Serve Knowledge Image",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/images/all/{md5}": {
      "get": {
        "description": "返回指定PDF的所有图片（单次请求，JSON + base64）",
        "operationId": "serve_batch_images_knowledge_images_all__md5__get",
        "parameters": [
          {
            "in": "path",
            "name": "md5",
            "required": true,
            "schema": {
              "title": "Md5",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Serve Batch Images",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/list": {
      "get": {
        "description": "获取用户的知识库文档列表",
        "operationId": "get_user_knowledge_list_knowledge_list_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KnowledgeListResponse_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Knowledge List",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/md5/clear": {
      "delete": {
        "description": "清空用户的MD5记录\n:param delete_documents: 是否同时删除知识库文档（默认True）",
        "operationId": "clear_user_md5_knowledge_md5_clear_delete",
        "parameters": [
          {
            "in": "query",
            "name": "delete_documents",
            "required": false,
            "schema": {
              "default": true,
              "title": "Delete Documents",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear User Md5",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/md5/delete/{md5_value}": {
      "delete": {
        "description": "删除单个MD5记录及其对应的知识库内容\n:param md5_value: 要删除的MD5值\n:param delete_documents: 是否同时删除知识库文档（默认True）",
        "operationId": "delete_single_md5_knowledge_md5_delete__md5_value__delete",
        "parameters": [
          {
            "in": "path",
            "name": "md5_value",
            "required": true,
            "schema": {
              "title": "Md5 Value",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "delete_documents",
            "required": false,
            "schema": {
              "default": true,
              "title": "Delete Documents",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Single Md5",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/md5/list": {
      "get": {
        "description": "获取用户的所有MD5记录",
        "operationId": "get_all_md5_records_knowledge_md5_list_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MD5ListResponse_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get All Md5 Records",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/knowledge/reranker/current": {
      "get": {
        "operationId": "get_current_reranker_config_knowledge_reranker_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current Reranker Config",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/knowledge/reranker/local-models": {
      "get": {
        "operationId": "list_local_reranker_models_knowledge_reranker_local_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Any_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Local Reranker Models",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}